\name{abduct.age.df}
\docType{data}
\alias{abduct.age.df}
\title{Victim abduct and the age data}
\description{
This data was studied to investigate whether there was a relationship between whether the victim had been abducted and the age of the victims in certain crimes. The age of the victims had been classiffed as 0-10 and 11+.
}
\usage{data(abduct.age.df)}
\references{
C. G. G. Aitken,  T. Connolly,  A. Gammerman,  G. Zhang,  D. Bailey, R. Gordon, and R. Oldfield. \emph{Statistical modelling in specific case analysis}. Science & Justice, 36(4):245-255, October 1996.
}
\keyword{datasets}