\name{anscombe.df}
\docType{data}
\alias{anscombe.df}
\title{Anscombe quartet}
\description{
This data is known as the Anscombe quartet. Each of sets has two variables, x and y. In each data set, x and y have the same mean (9 and 7.5), the same standard deviation (3.32 and 2.03) and the same correlation (0.82). 
}
\usage{data(anscombe.df)}
\references{
F. J. Anscombe. \emph{Graphs in statistical analysis}. The American Statistician, 27(1):17-21, 1973.
}
\keyword{datasets}