\name{bennett.df}
\docType{data}
\alias{bennett.df}
\title{Bennett data}
\description{
This data has 10 rows and 49 columns corresponding 10 refractive index (RI) measurements from 49 different locations in a windowpane.
}
\usage{data(bennett.df)}
\references{
R. L. Bennett, N. D. Kim, J. M. Curran, S. A. Coulson, and A. W. N. Newton. \emph{Spatial variation of refractive index in a pane of float glass}. Science & Justice, 43(2):71-76, April 2003.
}
\keyword{datasets}