\name{bottle.df}
\docType{data}
\alias{bottle.df}
\title{Bottle data}
\description{
This data contains the elemental concentration of five different elements (Manganese, Barium, Strontium, Zirconium, and Titanium) in samples of glass taken from six different Heineken beer bottles at four different locations (Base, Body, Shoulder, and Neck). Five repeat measurements are made on each sample at each location.
}
\usage{data(bottle.df)}
\references{
R. L. Bennett. \emph{Aspects of the analysis and interpretation of glass trace evidence}. Master's thesis, Department of Chemistry, University of Waikato, 2002.
}
\keyword{datasets}