\name{cck196.df}
\docType{data}
\alias{cck196.df}
\title{CCK-196G/A data}
\description{
This data has genotype of Cholecystokinin (CCK) gene promoter regions of -196G/A from selected suicide victims (S) and from control subjects (C).
}
\usage{data(cck196.df)}
\references{
Shindo S, Yoshioka N. \emph{Polymorphisms of the cholecystokinin gene promoter region in suicide victims in Japan}. Forensic Science International, 150(1):85-90, May 2005.
}
\keyword{datasets}