\name{cck45.df}
\docType{data}
\alias{cck45.df}
\title{CCK-45C/T data}
\description{
This data has genotype of Cholecystokinin (CCK) gene promoter regions of -45C/T from selected suicide victims (S) and from control subjects (C).
}
\usage{data(cck45.df)}
\references{
S. Shindo and N. Yoshioka. \emph{Polymorphisms of the cholecystokinin gene promoter region in suicide victims in Japan}. Forensic Science International, 150(1):85-90, May 2005.
}
\keyword{datasets}