\name{docexam.df}
\docType{data}
\alias{docexam.df}
\title{Misleading signatures data}
\description{
This data has number of misleading signatures in comparisons of 16 genuine signatures and 64 simulated signatures from 15 document examiners.
}
\usage{data(docexam.df)}
\references{
B. Found and D. K. Rogers. \emph{Investigating forensic document examiners' skill relating to opinions on photocopied signatures}. Science & Justice, 45(4):199-206, 2005.
}
\keyword{datasets}