\name{dpd.df}
\docType{data}
\alias{dpd.df}
\title{Deoxypyridinoline data}
\description{
This data measured deoxypyridinoline (DPD) to estimate human age. DPD is a nonreducible collagen crosslink that can be measured in human dentin samples extracted from permanent individual molars. Measurements were made in dentin samples from 22 patients with ages ranging from 15 to 73.
}
\usage{data(dpd.df)}
\references{
S. Martin-de las Heras, A. Valenzuela, and E. Villanueva. \emph{Deoxypyridinoline crosslinks in human dentin and estimation of age}. International Journal of Legal Medicine, 112(4):222-226, June 1999.
}
\keyword{datasets}