\name{fiber.color.df}
\docType{data}
\alias{fiber.color.df}
\title{Textile fibres in human hair data}
\description{
This data comes from a population study of textile fibres in human hair in Cambridgeshire, UK, 2002. It was carried out using 26 volunteers, and 12149 fibres were recovered from a variety of hair lengths using low adhesive tape and classified according to colour, generic type and fibre length. This data has percentage distribution of fibres in head hair according to colour.
}
\usage{data(fiber.color.df)}
\references{
R. Palmer and S. Oliver. \emph{The population of coloured fibres in human head hair}. Science & Justice, 44(2):83-88, April 2004.
}
\keyword{datasets}