\name{gc.df}
\docType{data}
\alias{gc.df}
\title{FBI-Gc data}
\description{
This data has the human group-specific component (Gc) genotypes of African Americans, Caucasians, and South Western Hispanics. The Gc locus has alleles A, B, and C and hence, the possible genotype are AA, AB, AC, BB, BC, and CC.
}
\usage{data(gc.df)}
\references{
FBI reference database
}
\keyword{datasets}