\name{ghb.df}
\docType{data}
\alias{ghb.df}
\title{Gamma-hydroxybutyric acid data}
\description{
This data comes from a gamma-hydroxybutyric acid (GHB) experiment, which has three groups of individuals:  (1) volunteers with no alcoholic history, (2) alcoholics who were waiting to start a GHB treatment program but had not yet received the drug, (3) alcoholics currently being treated with GHB, and measured the endogenous GHB concentration (ug/mL) in each.
}
\usage{data(ghb.df)}
\references{
F. Mari, L. Politi, C. Trignano, M. Grazia Di Milia, M. Di Padua, and E. Bertoli. \emph{What constitutes a normal ante-mortem urine GHB concentration?} Journal of Forensic and Legal Medicine, 16(3):148-151, April 2009.
}
\keyword{datasets}