\name{gustafson.df}
\docType{data}
\alias{gustafson.df}
\title{Teeth data}
\description{
This data is collected from 41 teeth, which were scored on a number of variables.
}
\usage{data(gustafson.df)}
\references{
G. Gustafson. \emph{Age determinations on teeth}.  Journal of the American Dental Association, 41(1):45-54, 1950.
}
\keyword{datasets}