\name{jauhari1.df}
\docType{data}
\alias{jauhari1.df}
\title{Pellet pattern data 1}
\description{
This data comes from a pilot experiment, which fired a sequence of 10 shots through a set of six paper targets that were equally spaced from 3 ft to 18 ft and recorded the size of the pellet pattern at 18 ft.  They then repeated the experiment with five of the targets removed and a single target at 18 ft. The third variable indicate with and without intermediate targets.
}
\usage{data(jauhari1.df)}
\references{
M. Jauhari, S. M. Chatterjee, and P. K. Ghosh. \emph{Statistical treatment of pellet dispersion data for estimating range of firing}. Journal of Forensic Sciences, 17(1):141-149, 1972.
}
\keyword{datasets}
