\name{jauhari2.df}
\docType{data}
\alias{jauhari2.df}
\title{Pellet pattern data 2}
\description{
This data comes from a pilot experiment by Jauhari et al. They fired a sequence of 10 shots through a set of six paper targets that were equally spaced from 3 ft to 18 ft and recorded the size of the pellet pattern at 18 ft.
}
\usage{data(jauhari2.df)}
\references{
M. Jauhari, S. M. Chatterjee, and P. K. Ghosh. \emph{Statistical treatment of pellet dispersion data for estimating range of firing}. Journal of Forensic Sciences, 17(1):141-149, 1972.
}
\keyword{datasets}
