\name{kerckring.df}
\docType{data}
\alias{kerckring.df}
\title{Kerckring data}
\description{
This data was analyzed the process of Kerckring on the occipital bone of numbers black and white perinates. The process of Kerckring is a projection of bone occasionally observed emerging from the inferior margin of the supraoccipital portion of the occipital squamous at the midline. Its status was regarded as either absent or present.
}
\usage{data(kerckring.df)}
\references{
S. M. Weinberg, D. A. Putz, M. P. Mooney, and M. I. Siegel. \emph{Evaluation of non-metric variation in the crania of black and white perinates}. Forensic Science International, 151(2-3):177-185, July 2005.
}
\keyword{datasets}