\name{last}
\alias{last}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Return the last element of a vector.
}
\description{
Return the last element of a vector
}
\usage{
last(x)
}
\arguments{
  \item{x}{
    a vector
}
}
\author{
James Curran
}
\examples{
x = 1:10
last(x)
}
