\name{liver.df}
\docType{data}
\alias{liver.df}
\title{Liver data}
\description{
The data record information on the presence or absence of extramedullary haematopoiesis (EMH) symptoms in 51 liver of sudden infant death (SIDS) and 102 non-SIDS cases.
}
\usage{data(liver.df)}
\references{
K. Toro, M.Hubay, and E. Keller. \emph{Extramedullary haematopoiesis in liver of sudden infant death cases}. Forensic Science International, 170(1):15-19, July 2007.
}
\keyword{datasets}