\name{morphine.df}
\alias{morphine.df}
\title{Morphine concentration in heart and perpipheral blood samples}
\description{
  Data was compiled from 126 morphine-involved cases
investigated by the Office of the Chief Medical Examiner, State
of Maryland, USA. An investigation was conducted into whether
comparison of morphine concentrations from a central and
peripheral site could be used to determine whether a morphine
death was acute or delayed. Fifty cases were identified as 'acute'
because the urine free morphine concentration by
radioimmunoassay (RIA) was less than 25 nglmL; 76 cases were
classified as 'random' because they had a urine morphine
concentration greater than 25 ng/mL by RIA. The average heart
blood to peripheral blood morphine concentration ratio in the
acute deaths was 1.40. The average heart blood to peripheral
blood morphine concentration ratio in the random deaths was
1.18. Because there was considerable overlap between the two
groups of data, the authors conclude that it was not possible to
predict 'acute' opiate intoxication deaths versus 'delayed' deaths
when the only information available is heart and peripheral
blood free morphine concentrations.
}
\usage{
data(morphine.df)
}
\references{
  B. Levine, D. Green-Johnson, K.A. Moore, D. Fowler, A. Jenkins, Assessment of the acuteness of heroin deaths from the analysis of multiple blood specimens, Science & Justice, Volume 42, Issue 1, January 2002, Pages 17-20.
}
\author{Levine et al.}
\keyword{datasets}
