\name{nasal.spline.df}
\docType{data}
\alias{nasal.spline.df}
\title{Nasal spine data}
\description{
This data was analyzed the projection of the anterior nasal spine on the maxillae of numbers black and white perinates. The anterior nasal spine was assessed by its forward projection away from the frontal plane of the anterior maxillary surface. This was best assessed from the lateral and/or superior perspective. Projection was described as slight, moderate or pronounced.
}
\usage{data(nasal.spline.df)}
\references{
S. M. Weinberg, D. A. Putz, M. P. Mooney, and M. I. Siegel. \emph{Evaluation of non-metric variation in the crania of black and white perinates}. Forensic Science International, 151(2-3):177-185, July 2005.
}
\keyword{datasets}