\name{newton.df}
\docType{data}
\alias{newton.df}
\title{Glass strata data}
\description{
This data has refractive index (RI) measurements for 30 fragments on each of the five strata - float surface (FS), near float surface (NFS), bulk (B), near anti-float (NAFS), and anti-float (AFS).
}
\usage{data(newton.df)}
\references{
A. W. N. Newton, J. M. Curran, C. M. Triggs, and J. S. Buckleton. \emph{The consequences of potentially differing distributions of the refractive indices of glass fragments from control and recovered sources}.  Forensic Science International, 140(2?3):185-193, March 2004.
}
\keyword{datasets}