\name{nzglass.df}
\docType{data}
\alias{nzglass.df}
\title{New Zealand glass RI data}
\description{
This data has refractive index (RI) measurements made on glass fragments recovered in New Zealand case work.
}
\usage{data(nzglass.df)}
\author{
John Buckleton, Sally Coulson, Tony Gummer, Angus Newton, Gerhard Wevers, Kevan Walsh (and ESR)
}
\keyword{datasets}