\name{palatal.df}
\docType{data}
\alias{palatal.df}
\title{Palatal arch shape data}
\description{
This data was analyzed the shape of the palatal arch on the maxillae of numbers black and white perinates. Palatal arch shape was determined by the curve formed by the inner alveolar margin. In order to determine shape, the separate maxillary halves were joined together at the midline. If the inner alveolar margins were observed to converge gradually along a continuous arc towards the midline, the palate was considered parabolic. On the other hand, if the inner alveolar margins were parallel to each other without converging until anterior of the premolars, the palate was considered hyperbolic. If the palate fell somewhere in between, it was considered intermediate.
}
\usage{data(palatal.df)}
\references{
S. M. Weinberg, D. A. Putz, M. P. Mooney, and M. I. Siegel. \emph{Evaluation of non-metric variation in the crania of black and white perinates}. Forensic Science International, 151(2-3):177-185, July 2005.
}
\keyword{datasets}