\name{ri.calibration.df}
\docType{data}
\alias{ri.calibration.df}
\title{Combined calibration data}
\description{
This data is combined from two calibration experiments which are conduct by Dr. Grzegorz Zadora and Bennett. The factor owner has two levels, RB for Rachel Bennett, and GZ for Grzegorz Zadora. The calibration of the instrument was established using standard glasses (Locke Scientific) set B1-B12 (RI=1.52912-1.520226). 
}
\usage{data(ri.calibration.df)}
\references{
Dr. Grzegorz Zadora from the Institute for Forensic Research in Krakow, Poland.
R. L. Bennett. \emph{Aspects of the analysis and interpretation of glass trace evidence}. Master's thesis, Department of Chemistry, University of Waikato, 2002.
}
\keyword{datasets}