\name{ri.calibration2.df}
\docType{data}
\alias{ri.calibration2.df}
\title{GRIM2 calibration data}
\description{
This data comes from a 2nd generation Glass Refractive Index Measurer (GRIM2) calibration experiment. The calibration of the instrument was established using standard glasses (Locke Scientific) set B1-B12 (RI=1.52912-1.520226). Each of the twelve reference sample glasses B1-B12 was measured five times.
}
\usage{data(ri.calibration2.df)}
\references{
Dr. Grzegorz Zadora from the Institute for Forensic Research in Krakow, Poland.
}
\keyword{datasets}