\name{sex.age.df}
\docType{data}
\alias{sex.age.df}
\title{Victims data}
\description{
This data is collected gender and age of victims in certain crimes. The age of the victims had been classified as 0-10 and 11+.
}
\usage{data(sex.age.df)}
\references{
C. G. G. Aitken, T. Connolly, A. Gammerman, G. Zhang, D. Bailey, R. Gordon, and R. Oldfield. \emph{Statistical modelling in specific case analysis}. Science & Justice, 36(4):245-255, October 1996.
}
\keyword{datasets}