\name{squamous.df}
\docType{data}
\alias{squamous.df}
\title{Occipital squamous bone data}
\description{
This data has been recorded the shape of the occipital squamous bone as narrow, equal, or greater for different races.
}
\usage{data(squamous.df)}
\references{
S. M. Weinberg, D. A. Putz, M. P. Mooney, and M. I. Siegel. \emph{Evaluation of non-metric variation in the crania of black and white perinates}. Forensic Science International, 151(2-3):177-185, July 2005.
}
\keyword{datasets}