\name{tryptase.df}
\docType{data}
\alias{tryptase.df}
\title{Tryptase data}
\description{
This data has tryptase concentrations, which was measured in blood from the femoral vein in 60 deaths: 39 control cases who died rapidly (within  minutes) from natural causes (sudden cardiac death and acute aortic dissection), 16 with death caused by prolonged asphyxia (traumatic compression of the chest and suffocation due to body position or smothering), and five anaphylactic deaths. In 44 of these cases, tryptase was measured in both heart (Tryp.cor) and femoral blood (Tryp.fem).
}
\usage{data(tryptase.df)}
\references{
E. Edston, O. Eriksson, M. V. Hage. \emph{Mast cell tryptase in postmortem serum - reference values and confounders}. International Journal of Legal Medicine, 121(4):275-280, May 2006
}
\keyword{datasets}