\name{velocity.df}
\docType{data}
\alias{velocity.df}
\title{Calibration data}
\description{
This data is about calibration between gun powder loading and speed.
}
\usage{data(velocity.df)}
\references{
S. C. K. Wong. \emph{The effects of projectile properties on glass backscatter: A statistical analysis}. Master's thesis, Department of Chemistry, University of Auckland, 2007. Forensic Science.
}
\keyword{datasets}