\name{wong.df}
\docType{data}
\alias{wong.df}
\title{Glass fragments data}
\description{
This data comes from a glass fragments experiment, count the number of glass fragments on the ground that were recovered after a window pane was shot with a handgun. The projectile velocity was controlled by altering the amount of gunpowder added to each bullet. The hardness (as measured on the Rockwell scale of hardness) of each projectile was altered by changing the amount of antimony (Sb) added to the projectile lead during casting. The profile of the projectile was changed by using a round-nose (RN) or wad-cutter (WC) mold. A full factorial design was used to allocate combinations of the factors (velocity, hardness, and profile) to the experimental units (shots). There were four velocity levels, three hardness levels, and two profile levels.
}
\usage{data(wong.df)}
\references{
S. C. K. Wong. \emph{The effects of projectile properties on glass backscatter: A statistical analysis}. Master's thesis, Department of Chemistry, University of Auckland, 2007. Forensic Science.
}
\keyword{datasets}